+++ commands/sparkles.py
@@ -26,7 +26,7 @@ class SparklesView(View):
     async def on_button_click(self, interaction):
         # Check if the user who clicked is the same as the one who initiated the command.
         if interaction.user.id != self.user_id:
-            await interaction.response.send_message("Only the user who typed the command can use these buttons.", ephemeral=True)
+            await interaction.response.send_message("Only the user who typed the command can use these buttons.😵‍💫 Use the commands instead.", ephemeral=True)
             return
         
         # Handle stop button click.
@@ -46,13 +46,14 @@ class SparklesCog(commands.Cog):
         self.bot = bot
         self.guild_states = {}  
         # List of sparkles sound file names and mapping to emoji labels
-        self.sparkles_sounds = ["sparkles00.mp3", "sparkles01.mp3", "sparkles02.mp3", "sparkles03.mp3"]
+        self.sparkles_sounds = ["sparkles00.mp3", "sparkles01.mp3", "sparkles02.mp3", "sparkles03.mp3", "sparkles04.mp3"]
         # Mapping of sound file names to emoji labels.
         self.sound_labels = {
             "sparkles00.mp3": "✨🪄⭐",
             "sparkles01.mp3": "✨🌟💫",
             "sparkles02.mp3": "✨🪄💎",
             "sparkles03.mp3": "✨🌲🌙",
+            "sparkles04.mp3": "✨🪄💫"
         }
 
     # Slash command to play sparkles sound.